Refactor ReadInput function into commonoperations package
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/dgschudde/advent-of-code-2024?shareId=XXXX-XXXX-XXXX-XXXX).
