+++ day-two/day-two.go
@@ -7,12 +7,13 @@ import (
 	"os"
 	"strconv"
 	"strings"
+	"commonoperations"
 )
 
 func main() {
 	var input *[]string
 
-	input = ReadInput()
+	input = commonoperations.ReadInput()
 
 	var totalSafeReports int
 
@@ -82,30 +83,4 @@ func checkSeries(series []int) string {
 		}
 	}
 	return "safe"
-}
-
-func ReadInput() *[]string {
-	var input = make([]string, 0)
-
-	// Read the input from file
-	inputFile, err := os.Open("./input/input.txt")
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	defer func(inputFile *os.File) {
-		err := inputFile.Close()
-		if err != nil {
-			log.Fatal(err)
-		}
-	}(inputFile)
-
-	scanner := bufio.NewScanner(inputFile)
-
-	for scanner.Scan() {
-		input = append(input, scanner.Text())
-	}
-
-	return &input
 }
